ai_scout_batch_2026_04_03/intelligent-rag-kg-fusion-prototype/requirements.txt
beautifulsoup4
chromadb
faiss-cpu
langchain
loguru
lxml
neo4j
numpy
openai
pandas
pydantic
pypdf
python-dotenv
requests
rdflib
scikit-learn
sentence-transformers
spacy
sqlalchemy
tiktoken
tqdm
transformers